+++ app/Http/Controllers/Api/ApiController.php
@@ -1,7 +1,9 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Api;
 
+use App\Http\Controllers\Support\UsesJsonResponse;
+use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Foundation\Bus\DispatchesJobs;
 use Illuminate\Routing\Controller as BaseController;
 use Illuminate\Foundation\Validation\ValidatesRequests;
@@ -9,5 +11,5 @@ use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class ApiController extends BaseController
 {
-    use AuthorizesRequests, DispatchesJobs, ValidatesRequests;
+    use AuthorizesRequests, DispatchesJobs, SoftDeletes, UsesJsonResponse, ValidatesRequests;
 }
