+++ app/Http/Controllers/Api/CategoryController.php
@@ -1,17 +1,14 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Api;
 
-use App\Http\Requests\Tag\StoreRequest;
-use App\Http\Requests\Tag\UpdateRequest;
-use App\Models\Tag;
-use Illuminate\Database\Eloquent\SoftDeletes;
+use App\Http\Requests\Category\StoreRequest;
+use App\Http\Requests\Category\UpdateRequest;
+use App\Models\Category;
 use Illuminate\Http\Response;
 
-class TagController extends Controller
+class CategoryController extends ApiController
 {
-    use SoftDeletes;
-
     /**
      * Display a listing of the resource.
      *
@@ -19,7 +16,7 @@ class TagController extends Controller
      */
     public function index()
     {
-        return $this->sendJson(Tag::paginate());
+        return $this->sendJson(Category::paginate());
     }
 
     /**
@@ -40,31 +37,31 @@ class TagController extends Controller
      */
     public function store(StoreRequest $request)
     {
-        $tag = new Tag();
-        $tag->fill($request->all());
-        $tag->save();
+        $category = new Category();
+        $category->fill($request->all());
+        $category->save();
 
-        return $this->sendJson($tag, Response::HTTP_CREATED);
+        return $this->sendJson($category, Response::HTTP_CREATED);
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Tag  $tag
+     * @param  \App\Models\Category  $category
      * @return \Illuminate\Http\Response
      */
-    public function show(Tag $tag)
+    public function show(Category $category)
     {
-        return $this->sendJson($tag);
+        return $this->sendJson($category);
     }
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Models\Tag  $tag
+     * @param  \App\Models\Category  $category
      * @return \Illuminate\Http\Response
      */
-    public function edit(Tag $tag)
+    public function edit(Category $category)
     {
         //
     }
@@ -73,26 +70,25 @@ class TagController extends Controller
      * Update the specified resource in storage.
      *
      * @param  UpdateRequest  $request
-     * @param  \App\Models\Tag  $tag
-     *
+     * @param  \App\Models\Category  $category
      * @return \Illuminate\Http\Response
      */
-    public function update(UpdateRequest $request, Tag $tag)
+    public function update(UpdateRequest $request, Category $category)
     {
-        $tag->update($request->all());
+        $category->update($request->all());
 
-        return $this->sendJson($tag, Response::HTTP_OK);
+        return $this->sendJson($category, Response::HTTP_OK);
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Tag  $tag
+     * @param  \App\Models\Category  $category
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Tag $tag)
+    public function destroy(Category $category)
     {
-        $tag->delete();
+        $category->delete();
 
         return $this->sendJson([], Response::HTTP_NO_CONTENT);
     }
