+++ app/Http/Controllers/Api/TagController.php
@@ -1,17 +1,14 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Api;
 
-use App\Http\Requests\Category\StoreRequest;
-use App\Http\Requests\Category\UpdateRequest;
-use App\Models\Category;
-use Illuminate\Database\Eloquent\SoftDeletes;
+use App\Http\Requests\Tag\StoreRequest;
+use App\Http\Requests\Tag\UpdateRequest;
+use App\Models\Tag;
 use Illuminate\Http\Response;
 
-class CategoryController extends Controller
+class TagController extends ApiController
 {
-    use SoftDeletes;
-
     /**
      * Display a listing of the resource.
      *
@@ -19,7 +16,7 @@ class CategoryController extends Controller
      */
     public function index()
     {
-        return $this->sendJson(Category::paginate());
+        return $this->sendJson(Tag::paginate());
     }
 
     /**
@@ -40,31 +37,31 @@ class CategoryController extends Controller
      */
     public function store(StoreRequest $request)
     {
-        $category = new Category();
-        $category->fill($request->all());
-        $category->save();
+        $tag = new Tag();
+        $tag->fill($request->all());
+        $tag->save();
 
-        return $this->sendJson($category, Response::HTTP_CREATED);
+        return $this->sendJson($tag, Response::HTTP_CREATED);
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Category  $category
+     * @param  \App\Models\Tag  $tag
      * @return \Illuminate\Http\Response
      */
-    public function show(Category $category)
+    public function show(Tag $tag)
     {
-        return $this->sendJson($category);
+        return $this->sendJson($tag);
     }
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Models\Category  $category
+     * @param  \App\Models\Tag  $tag
      * @return \Illuminate\Http\Response
      */
-    public function edit(Category $category)
+    public function edit(Tag $tag)
     {
         //
     }
@@ -73,25 +70,26 @@ class CategoryController extends Controller
      * Update the specified resource in storage.
      *
      * @param  UpdateRequest  $request
-     * @param  \App\Models\Category  $category
+     * @param  \App\Models\Tag  $tag
+     *
      * @return \Illuminate\Http\Response
      */
-    public function update(UpdateRequest $request, Category $category)
+    public function update(UpdateRequest $request, Tag $tag)
     {
-        $category->update($request->all());
+        $tag->update($request->all());
 
-        return $this->sendJson($category, Response::HTTP_OK);
+        return $this->sendJson($tag, Response::HTTP_OK);
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Category  $category
+     * @param  \App\Models\Tag  $tag
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Category $category)
+    public function destroy(Tag $tag)
     {
-        $category->delete();
+        $tag->delete();
 
         return $this->sendJson([], Response::HTTP_NO_CONTENT);
     }
